requirements.txt
Django~=5.0
psycopg2-binary~=2.9.9
django-ipware~=6.0.3
factory-boy~=3.3.0